alm-optimization-ml/requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
matplotlib
streamlit
